+++ src/CMakeLists.txt
@@ -567,6 +567,8 @@ list( APPEND SOURCES
          effects/audiounits/AUControl.mm
          effects/audiounits/AudioUnitEffect.cpp
          effects/audiounits/AudioUnitEffect.h
+         effects/audiounits/AudioUnitEffectOptionsDialog.cpp
+         effects/audiounits/AudioUnitEffectOptionsDialog.h
          effects/audiounits/AudioUnitEffectsModule.cpp
          effects/audiounits/AudioUnitEffectsModule.h
          effects/audiounits/AudioUnitUtils.cpp
